Fix build
It's not good to use an absolute path here because the version of /usr/bin/python is > 2.7 for most of us I guess.
So I put a link to python2 into /Android/bin and added it to the beginning of my PATH variable. But this link isn't used when using an absolute path here.